Remove clearing in PEA phase.

+++ dataset/GitHub_Java/oracle.graal/compiler/src/org.graalvm.compiler.truffle.compiler/src/org/graalvm/compiler/truffle/compiler/nodes/frame/VirtualFrameFreeNode.java
@@ -48,15 +48,16 @@ public class VirtualFrameFreeNode extends VirtualFrameAccessorNode implements Vi
     @Override
     public void virtualize(VirtualizerTool tool) {
         ValueNode tagAlias = tool.getAlias(frame.virtualFrameTagArray);
-        if (tagAlias instanceof VirtualObjectNode) {
+        ValueNode objectStorageAlias = tool.getAlias(frame.virtualFrameObjectArray);
+        ValueNode primitiveStorageAlias = tool.getAlias(frame.virtualFramePrimitiveArray);
+
+        if (tagAlias instanceof VirtualObjectNode && objectStorageAlias instanceof VirtualObjectNode && primitiveStorageAlias instanceof VirtualObjectNode) {
             VirtualObjectNode tagVirtual = (VirtualObjectNode) tagAlias;
-            if (frameSlotIndex < tagVirtual.entryCount()) {
-                // Simply set kind to illegal. A later phase will clear the slots.
-                boolean success = tool.setVirtualEntry(tagVirtual,
-                                frameSlotIndex,
-                                getConstant(accessTag),
-                                tagVirtual.entryKind(tool.getMetaAccessExtensionProvider(), frameSlotIndex),
-                                -1);
+            VirtualObjectNode objectStorageVirtual = (VirtualObjectNode) objectStorageAlias;
+            VirtualObjectNode primitiveStorageVirtual = (VirtualObjectNode) primitiveStorageAlias;
+
+            if (frameSlotIndex < tagVirtual.entryCount() && frameSlotIndex < objectStorageVirtual.entryCount() && frameSlotIndex < primitiveStorageVirtual.entryCount()) {
+                boolean success = tool.setVirtualEntry(tagVirtual, frameSlotIndex, getConstant(accessTag), tagVirtual.entryKind(tool.getMetaAccessExtensionProvider(), frameSlotIndex), -1);
                 if (success) {
                     tool.delete();
                     return;
